+++ tests/testthat/test-files.R
@@ -77,7 +77,7 @@ test_that("scoped backups are cleared as necessary", {
   list.files(tempdir())
 
   files <- list.files(tempdir())
-  backup <- grep("^renv-backup-", files)
+  backup <- grep("^\\.renv-backup-", files)
   expect_length(backup, 0)
 
 })
@@ -124,5 +124,34 @@ test_that("attempts to overwrite existing files are handled appropriately", {
 
   expect_error(renv_file_copy(source, target))
   expect_true(renv_file_copy(source, target, overwrite = TRUE))
+
+})
+
+test_that("permissions, timestamps are preserved", {
+
+  source <- renv_tempfile("renv-source-")
+  target <- renv_tempfile("renv-target-")
+
+  ensure_directory(source)
+
+  range <- 1:10
+  files <- sprintf("%02i.txt", range)
+  for (i in range) {
+    Sys.sleep(0.01)
+    file.create(file.path(source, sprintf("%02i.txt", i)))
+  }
+
+  renv_file_copy(source, target)
+
+  srcfiles <- list.files(source, full.names = TRUE)
+  tgtfiles <- list.files(target, full.names = TRUE)
+
+  srcinfo <- file.info(srcfiles)
+  tgtinfo <- file.info(tgtfiles)
+
+  rownames(srcinfo) <- rownames(tgtinfo) <- basename(srcfiles)
+
+  fields <- setdiff(names(srcinfo), c("ctime", "atime"))
+  expect_equal(srcinfo[fields], tgtinfo[fields])
 
 })